Added skip probability to room colours
Was thinking that you could tell how many people had been in the room
since it had been regenerated if you counted the colours, which kinda
leaks data. Now it could skip up to 2 colours, which is less leaky, but
we since we're not spiralling into darkness with colours anymore we
could probably push this further.