Created home page, updated server to match
Going to / no longer redirs to /-, and instead there is a snazzy home
page. The rooms page is moved to rooms.html from index.html, and the
index.html in it's place is the home page.

I've also added a rudimentary impl. of "random room" button, and moved
the idea of /- to /all as that makes sense to me.